refactor: split Dagger module into per-component objects + ADR0013 domain types
Replaces the flat SkeOperatorBuild/KratixUnit/... function soup with typed,
chainable Dagger objects: SkeOperator, Kratix, KratixCli, one file each,
constructed via Ci.SkeOperator(source)/.Kratix(source)/.KratixCli(source).
CLI shape changes from 'ske-operator-build' to
'ske-operator --source=... build', matching Dagger's own object-chaining
idiom instead of name-prefixed flat functions.

All three satisfy a shared Pipeline interface (Unit + IsReleasable), so the
compiler enforces the same shape rather than relying on convention. Build
and SystemTest stay concrete per component since their signatures genuinely
differ (Container vs File artifact, tokens or not).

Adds the ADR0013 domain model for real: Component, Artifact, Release,
SkeRelease, KratixRelease in release.go. Each component's IsReleasable now
dispatches through a Release value instead of returning a bare stubbed bool.
Also adds Commit and Tag (not in ADR0013) so the commit -> tag -> release
path has real types; Release.Tag stays zero-value until phase 2 plumbs a
commit SHA into these calls.

Found and fixed along the way: the source directory field must be exported
(Source, not source) for Dagger to persist it as part of the object's state
between calls — an unexported field comes back nil on the hydrated object,
causing a panic on the first chained call.

Re-verified ske-operator-build (now 'ske-operator ... build') and
is-releasable for all three components after the restructure.
